pyproject.toml
[project]
name = "ml-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fugashi>=1.4.0",
    "hydra-core>=1.3.2",
    "lightning-utilities>=0.14.0",
    "mecab-python3>=1.0.10",
    "mlflow>=1.0.0",
    "numpy<2",
    "polars>=1.24.0",
    "pytorch-lightning>=2.5.0.post0",
    "rich>=13.9.4",
    "rootutils>=1.0.7",
    "torch==2.2",
    "torchvision>=0.17.0",
    "transformers>=4.49.0",
    "unidic-lite>=1.0.8",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
